renaming Voyage components
