+++ src/components/TravelCreationPopup/index.js
@@ -1,0 +1,3 @@
+import TCPopup from './TravelCreationPopup';
+
+export default TCPopup;
